move insert mutation to Returns
