+++ client/packages/invoices/src/InboundShipment/api/api.ts
@@ -13,7 +13,6 @@ import {
   Formatter,
   UpdateInboundShipmentStatusInput,
   setNullableInput,
-  SupplierReturnInput,
 } from '@openmsupply-client/common';
 import { DraftInboundLine } from './../../types';
 import { isA } from '../../utils';
@@ -359,12 +358,5 @@ export const getInboundQueries = (sdk: Sdk, storeId: string) => ({
     }
 
     throw new Error('Could not add from master list');
-  },
-  insertSupplierReturn: async (input: SupplierReturnInput) => {
-    const result = await sdk.insertSupplierReturn({
-      input,
-    });
-
-    return result.insertSupplierReturn;
   },
 });
