+++ client/packages/invoices/src/InboundShipment/api/hooks/index.ts
@@ -13,7 +13,6 @@ export const useInbound = {
     insert: Document.useInsertInbound,
     update: Document.useUpdateInbound,
     updateTax: Document.useUpdateInboundServiceTax,
-    insertSupplierReturn: Document.useInsertSupplierReturn,
 
     fields: Document.useInboundFields,
     next: Document.useNextItem,
