+++ client/packages/invoices/src/InboundShipment/api/operations.graphql
@@ -639,13 +639,3 @@ mutation addToInboundShipmentFromMasterList(
     }
   }
 }
-
-mutation insertSupplierReturn($input: SupplierReturnInput!) {
-  insertSupplierReturn(input: $input) {
-    ... on InvoiceNode {
-      __typename
-      id
-      invoiceNumber
-    }
-  }
-}
