+++ client/packages/invoices/src/Returns/api/hooks/document/index.ts
@@ -1,0 +1,5 @@
+import { useInsertSupplierReturn } from './useInsertSupplierReturn';
+
+export const Document = {
+  useInsertSupplierReturn,
+};
