+++ client/packages/invoices/src/Returns/api/hooks/index.ts
@@ -1,7 +1,11 @@
+import { Document } from './document';
 import { Lines } from './line';
 import { Utils } from './utils';
 
 export const useReturns = {
+  document: {
+    insertSupplierReturn: Document.useInsertSupplierReturn,
+  },
   lines: {
     newReturnLines: Lines.useNewSupplierReturnLines,
   },
