+++ client/packages/invoices/src/Returns/api/operations.graphql
@@ -17,3 +17,13 @@ query newSupplierReturnLines(
     stockLineId
   }
 }
+
+mutation insertSupplierReturn($input: SupplierReturnInput!) {
+  insertSupplierReturn(input: $input) {
+    ... on InvoiceNode {
+      __typename
+      id
+      invoiceNumber
+    }
+  }
+}
